🐛 refactor(tests): remove skipAllSleeps from FakeSleeper
The `skipAllSleeps` functionality in `FakeSleeper` has been removed. This change simplifies the fake implementation and enforces a more realistic testing approach by requiring explicit time advancement. Tests that relied on skipping sleeps will now need to be updated to use `advanceTime` to simulate the passage of time. This leads to more robust and accurate test scenarios.