✅ refactor(errors): remove unnecessary await from alert calls (#70)
The `alert` function in the codebase is not an asynchronous function, therefore it does not need to be `await`ed. This commit removes the unnecessary `await` keywords from all calls to `alert` across various ViewModels. This change simplifies the code and ensures that the UI can be updated correctly and synchronously when errors are presented to the user.

Specifically, the following files have been updated:
- `EpisodeDetailViewModel.swift`
- `EpisodesListViewModel.swift`
- `PodcastDetailViewModel.swift`
- `SelectablePodcastsGridViewModel.swift`
- `OPMLViewModel.swift`
- `UpNextViewModel.swift`